7. 모아보기 테스트 이동
