+++ src/test/java/com/adventours/calendar/gift/GiftControllerTest.java
@@ -1,5 +1,6 @@
 package com.adventours.calendar.gift;
 
+import com.adventours.calendar.auth.JwtTokenIssuer;
 import com.adventours.calendar.calendar.domain.Calendar;
 import com.adventours.calendar.calendar.persistence.CalendarRepository;
 import com.adventours.calendar.common.ApiTest;
@@ -16,8 +17,13 @@ import io.restassured.RestAssured;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.mockito.BDDMockito;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.SpyBean;
 
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneOffset;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -33,6 +39,10 @@ class GiftControllerTest extends ApiTest {
     @Autowired
     GiftService giftService;
     @Autowired UserRepository userRepository;
+    @Autowired
+    JwtTokenIssuer jwtTokenIssuer;
+    @SpyBean
+    Clock clock;
 
     @Test
     @DisplayName("선물 생성/수정 성공")
@@ -79,12 +89,15 @@ class GiftControllerTest extends ApiTest {
                 .statusCode(200);
     }
 
+
     @Test
-    @DisplayName("내가 발행한 캘린더의 선물 summary")
-    void getMyGiftListSummaryList() {
-        Scenario.createCalendar().request()
-                .updateGift().request();
-        final Calendar calendar = calendarRepository.findAll().get(0);
+    @DisplayName("내 캘린더 선물 모아보기")
+    void getSummaryMyCalendarList() {
+        final User user = Scenario.createUserDB().id(1L).create();
+        final Calendar calendar = Scenario.createCalendarDB().uuid(UUID.randomUUID()).user(user).create();
+        Scenario.updateGift().giftId(1L).request()
+                .updateGift().giftId(2L).request()
+                .updateGift().giftId(3L).request();
 
         RestAssured.given().log().all()
                 .header("Authorization", accessToken)
@@ -96,16 +109,27 @@ class GiftControllerTest extends ApiTest {
     }
 
     @Test
-    @DisplayName("구독한 선물 열기 성공")
-    void openGift_sub() {
-        final User user = Scenario.createUserDB().id(2L).create();
-        final Calendar calendar = Scenario.createCalendarDB().uuid(UUID.randomUUID()).user(user).create();
-        Scenario.subscribeCalendar().calendarId(calendar.getId()).request().
-                openGift().calendarId(calendar.getTitle()).request();
+    @DisplayName("구독 캘린더 선물 모아보기 - 열람 가능한 선물 모아보기")
+    void getSummarySubCalendarList() {
+        final User user2 = Scenario.createUserDB().id(2L).create();
+        final String user2AccessToken = jwtTokenIssuer.issueToken(user2.getId()).accessToken();
+        final Calendar calendar = Scenario.createCalendarDB().uuid(UUID.randomUUID()).user(user2).create();
+        Scenario.updateGift().accessToken(user2AccessToken).giftId(1L).request()
+                .updateGift().accessToken(user2AccessToken).giftId(2L).request()
+                .updateGift().accessToken(user2AccessToken).giftId(3L).request();
 
-        final GiftPersonalState giftPersonalState = giftPersonalStateRepository.findAll().get(0);
 
-        assertThat(giftPersonalState.isOpened()).isTrue();
+        Instant instant = Instant.now().atZone(ZoneOffset.UTC).plusHours(9)
+                .withMonth(12).withDayOfMonth(13).toInstant();
+        BDDMockito.given(clock.instant())
+                .willReturn(instant);
+        RestAssured.given().log().all()
+                .header("Authorization", accessToken)
+                .when()
+                .get("/calendar/{calendarId}/gift/sub/summary", calendar.getId())
+                .then()
+                .log().all()
+                .statusCode(200);
     }
 
     @Test
